Add now ls support.
